Made 42kit have its own CMakeLists.txt for more clear separation and some independ built rules, if desired. This required moving around some graphics stuff. Also added address sanitizer (asan) as an option for DEBUG builds (simply, its a very nice memory analysis tool). Next task, got bug hunting with asan.

+++ Source/42init.c
@@ -1307,7 +1307,7 @@ void InitRigidDyn(struct SCType *S)
    /* Body Path Table */
    /* Body in path -> 1, else 0 */
    D->BodyPathTable = (struct BodyPathTableType **)calloc(
-       S->Nb, sizeof(struct BodyPathTableType *));
+       (int)S->Nb, sizeof(struct BodyPathTableType *));
    for (i = 0; i < S->Nb; i++)
       D->BodyPathTable[i] = (struct BodyPathTableType *)calloc(
           S->Nb, sizeof(struct BodyPathTableType));
